add cmd exec logic to loop. returns output. see comment.
handlers for command execution added. Calls the package wrappers we made for command execution. output will be returned for non-interactive commands. output will be displayed on the viewport for the user. Eventually, errors should be logged (they are not logged or printed). There are also some weird quirks with the key reading. key commands will be read by the viewport no matter which model is selected by the user. Also, flags are currently ignored...